feat(parquet): relocate arrow coercion and sbbf check
